move self signed certificate method to own file
